+++ test/dartemis/core/utils/_all_tests.dart
@@ -1,7 +1,9 @@
 library utils._all_tests;
 
 import 'bag_test.dart' as bag;
+import 'entity_bag_test.dart' as entity_bag;
 
 void main() {
   bag.main();
+  entity_bag.main();
 }