+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * renamed `Poolable` to `Pooled`
 * renamed `ComponentPoolable` to `PooledComponent`
 * removed `FastMath` and `Utils`, unrelated to ECS
+* removed `removeAll` from `Bag`
 * `time` and `frame` getters have been moved from `World` to `EntitySystem`, `World` has methods instead
 ### API Changes
 * deprecated `ComponentMapper` use `Mapper` instead
@@ -12,6 +13,7 @@
 * `System`s can be assigned to a group when adding them to the `World`, `Word.process()` can be called for a specific group
 ### Enhancements
 * created a code generating transformer to insert `Manager`, `System`s and `Mapper`s into `Manager` and `System` instances
+* performance improvement when removing entities
 ### Bugfixes
 * DelayedEntityProcessingSystem keeps running until all current entities have expired
 ### Internal